requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2020.6.20
chardet==4.0.0
cryptography==3.3.2
decorator==5.1.1
distro==1.5.0
exceptiongroup==1.1.3
executing==2.0.0
idna==2.10
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
pyOpenSSL==20.0.1
reportlab==3.5.59
requests==2.25.1
six==1.16.0
stack-data==0.6.3
traitlets==5.11.2
typing-extensions==4.8.0
urllib3==1.26.5
wcwidth==0.2.8
request-manager-attenuare
tqdm
